Trading-Project/requirements.txt
# Python Dependencies
python-dotenv
requests
pandas
schedule
scipy
matplotlib
numpy
bandit